[feature] adjust get notes, add detail note, add delete note

+++ src/main/java/willy/project250630/dto/notes/NoteDto/ResNoteDto.java
@@ -1,10 +1,11 @@
 package willy.project250630.dto.notes.NoteDto;
 
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
 @Data
 public class ResNoteDto {
-    private Integer id;
+    private Long id;
     private String title;
     private String content;
 }
